Use django.contrib.staticfiles
This is the default Django way of serving static files.

For deployment, the collectstatic command can be used to collect all static
files into one directory and serve it using a 'real' webserver.

In runserver mode, static files for django admin apps as well as custom
apps are automatically served by the included webserver.

Conflicts:
	src/templates/attestation/attestation_edit.html
	src/templates/attestation/attestation_list.html
	src/templates/base.html
	src/templates/solutions/checker_result_list.html

Conflicts:
	src/settings/base.py
	src/settings/devel.py
	src/settings/local.py-dist
	src/templates/attestation/attestation_edit.html
	src/templates/attestation/attestation_list.html
	src/templates/base.html
	src/templates/tasks/task_list.html