Implement test uploads
These are solution uploads that are otherwise ignored (possibly not
complete yet, but in some place at least) and only available to tutors
and trainers. They are also not restricted by deadlines. Links to the
submission pages are on the admin view. (TODO: Where to link to them for
tutors?)

These are _not_ removed. I figured it is useful to keep them around, in
case a tutor wants to show the results to a trainer or something.

This closes: #210