Initial support for jPlag integration
TODO: Hide links if JPLAGJAR ist not set (i.e. there is no jplag
support), and gracfully abort the view.

Add a field to the task database (jplag up-to-date), that is set when
jPlag is run, and that is cleared whenever a new final solution is
uploaded. Display that field.

Maybe make running jplag a bit more robust against simulatenous runs,
i.e. work in a temporary directory and move over at the end.