regular expression and Compiler C
When we select the compiler we already know that it is a necessary
requirement. A program with an error does not compile.
That is why I have made this small modification so ‘Required’ is
automatically selected in the Compiler C.

I want to suggest to implement the regular expression:

“^[a-zA-Z0-9_ . -\?!]*.[cC]$”

As a possible side effect, the student that is used to other IDEs in
which the same solution do not represent a error source, e.g. Visual
studio, is discouraged to use Praktomat.