Text Checker
You can choose what it has be or it has not to be in the student's
solution
Example:
In the C language it is not recommended to use the "goto". You can
select "The text must not be in the solution" and insert in the Text
"goto". Text Checker will search "goto" in the student's solution. If it
finds the text, depending on the basis of your selection (Critical or
non-Critical) it gives back an error or a warning.On the contrary, you
can select "The text has to be in the solution" and Text Checker will
check whether the Text you inserted is in the solution.