Only search for main entry point if required
For the JAVA builder the process of determining the main class, is
quite time consuming. For example, in the "Database System Architecture
and Implementation" [1] searching for a main entry point in the long
list of java classes caused server timeouts.

Thus, only search for a main entry point, if that is required.

[1] https://cms.uni-konstanz.de/informatik/grossniklaus/education/archive/inf-20210/