Switch to GitHub Actions for running tests
As we don’t need the Travis CI anymore, the corresponding build configuration file is also going to be removed.

gjc-jdk won’t be installed anymore because it is not available anymore in recent versions of Ubuntu. However, this is not a problem since the tests don’t rely on the presence of the package.

Everything else should work as it did before.