safeexec: replace preexec_fn
Since Python 3.8, the preexec_fn parameter (subprocess.Popen) is not supported within subinterpreters anymore. Generally, you're also advised to not use this paramater if possible (outside of subinterpreters).

Instead of using a preexec function, this patch makes use of the start_new_session parameter to set the session id. The prlimit tool is now used to limit the resources (number of open file descriptors and maximum file size).